% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACTS-package.R
\docType{package}
\name{BACTS-package}
\alias{BACTS-package}
\alias{BACTS}
\title{The 'BACTS' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
