% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hprob.R
\name{hprob}
\alias{hprob}
\title{Weight function based on the similarity probability of two samples}
\usage{
hprob(k = 1, lambda = 0.4, p)
}
\arguments{
\item{k}{scalar (positive) ; parameter controlling the shape of the function.}

\item{lambda}{scalar (positive - between 0 and 1) ; parameter controlling the scale of the function.}

\item{p}{scalar (between 0 and 1) ; probability to get weight for - here this is related on how similar two posterior samples appear to be.}
}
\description{
This function provides a weight with respect to the probability 'p' describing the similarity of two (posterior) samples. Its main feature is the symmetry around .5 and it is based on two parameters 'lamda' and 'k', controlling the scale and shape respectively, determining the penalty level as 'p' deviates for .5.
}
\details{
The choice of parameters 'lambda' and 'k' denote the scale and shape of the function respectively. Specifically, larger values of 'lambda' and/or 'k' tend to produce a faster decrease from the peak of the function, whereas, smaller values correspond to a smoother and slower decrease from the peak.
}
\examples{
hprob(k = 1, lambda = .1, p = .1)

}
